requirements.txt
asn1crypto==0.24.0
braintree==3.45.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==2.0.5
django-payments==0.12.3
idna==2.6
pycparser==2.18
PyJWT==1.6.1
pytz==2018.4
rauth==0.7.3
requests==2.18.4
six==1.11.0
stripe==1.80.0
suds-jurko==0.6
todoist-python==7.0.18
urllib3==1.22
xmltodict==0.11.0
